requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.2
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
pandas-ta>=0.3.14b0
yfinance>=0.2.40
httpx>=0.27.0
openai>=1.43.0
cachetools>=5.4.0
